refactor: domain 의 CalculateRepository 를 이용하도록 변경
