+++ calculator/build.gradle
@@ -39,6 +39,15 @@ android {
         pickFirst 'META-INF/AL2.0'
         pickFirst 'META-INF/LGPL2.1'
     }
+    sourceSets {
+        final String SHARED_TEST_DIR = 'src/testShared/java'
+        test {
+            java.srcDirs += SHARED_TEST_DIR
+        }
+        androidTest {
+            java.srcDirs += SHARED_TEST_DIR
+        }
+    }
 }
 
 dependencies {
